requirements.txt
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-bootstrap4==0.0.7
-e git+https://github.com/tykling/django-eav2.git@992984bb4b2725293b1029f78f1ecf97924a80bc#egg=django_eav2
factory-boy==2.11.1
Faker==0.9.2
idna==2.7
oauthlib==2.1.0
psycopg2-binary==2.7.5
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
requests==2.20.0
requests-oauthlib==1.0.0
six==1.11.0
text-unidecode==1.2
urllib3==1.24
wrapt==1.10.11
